clinic review status update for protocol
