+++ streamlit_apps/mert_components/protocol_review.py
@@ -8,20 +8,7 @@ from datetime import datetime
 import pandas as pd
 import streamlit as st
 
-mert2_user = {
-    "STF-e465eb68-ba87-11eb-8611-06b700432873": "Luis Camargo",
-    "STF-6d38ac86-ba89-11eb-8b42-029e69ddbc8b": "Alex Ring",
-    "STF-ac677ad4-a595-11ec-82d9-02fd9bf033d7": "Stephanie Llaga",
-    "STF-e8b6c0a2-27f5-11ed-b837-02b0e344b06f": "Patrick Polk",
-    "STF-934d6632-a17e-11ec-b364-0aa26dca46cb": "Joseph Chong",
-    "STF-031845e2-b505-11ec-8b5d-0a86265d54df": "Nicole Yu",
-    "STF-d844feb2-241c-11ef-8e46-02fb253d52c7": "Binh Le",
-    "STF-7a0aa2d4-241c-11ef-a5ac-06026d518b71": "Rey Mendoza",
-    "STF-0710bc38-2e40-11ed-a807-027d8017651d": "Jay Kumar",
-    "STF-472808de-ba89-11eb-967d-029e69ddbc8b": "Jijeong Kim",
-    "STF-143c1a12-8657-11ef-8e6a-020ab1ebdc67": "Uma Gokhale",
-    "STF-8b4db98a-8657-11ef-9d8b-020ab1ebdc67": "Uma Gokhale2",
-}
+from .review_utils import EEGReviewState, mert2_user
 
 
 @st.fragment
@@ -33,6 +20,12 @@ def render_protocol_page(data_manager):
     base_protocol = eeg_info["baseProtocol"]
     analysis_meta = eeg_info["eegInfo"]["analysisMeta"]
     eeg_info_data = eeg_info["eegInfo"]
+
+    current_state = (
+        EEGReviewState[analysis_meta["reviewState"]]
+        if analysis_meta["reviewState"]
+        else EEGReviewState.PENDING
+    )
 
     if "protocol" in eeg_info:
         protocol_data = eeg_info["protocol"]
@@ -222,6 +215,12 @@ def render_protocol_page(data_manager):
                     # Ran twice to complete, protocol review by eeg scientist
                     asyncio.run(data_manager.save_protocol(protocol))
 
+                    asyncio.run(
+                        data_manager.update_eeg_review(
+                            is_first_reviewer=(current_state == EEGReviewState.PENDING),
+                            state=EEGReviewState.CLINIC_REVIEW.name,
+                        )
+                    )
                     st.success("Protocol updated successfully!")
                 except Exception as e:
                     st.error(f"Failed to update protocol: {str(e)}")
